requirements.txt
streamlit
supabase
python-dotenv
psycopg2-binary
psycopg2
cohere